HIR-32 refactor: 비즈니스 로직을 서비스 계층으로 이동함
