+++ src/main/java/kr/binarybard/hireo/api/auth/controller/AuthenticationApiController.java
@@ -1,20 +1,13 @@
 package kr.binarybard.hireo.api.auth.controller;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.Valid;
 import kr.binarybard.hireo.api.auth.dto.RefreshTokenRequest;
 import kr.binarybard.hireo.api.auth.dto.SignInRequest;
-import kr.binarybard.hireo.api.auth.service.RefreshTokenService;
-import kr.binarybard.hireo.common.exceptions.AuthException;
-import kr.binarybard.hireo.common.exceptions.ErrorCode;
-import kr.binarybard.hireo.config.jwt.JwtTokenProvider;
-import kr.binarybard.hireo.web.auth.service.LoginService;
+import kr.binarybard.hireo.api.auth.dto.TokenResponse;
+import kr.binarybard.hireo.api.auth.service.JwtLoginService;
+import kr.binarybard.hireo.api.auth.service.JwtTokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,63 +17,16 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/auth")
 @RequiredArgsConstructor
 public class AuthenticationApiController {
-	private final JwtTokenProvider tokenProvider;
-	private final AuthenticationManagerBuilder authenticationManagerBuilder;
-	private final LoginService loginService;
-	private final RefreshTokenService refreshTokenService;
+	private final JwtLoginService loginService;
+	private final JwtTokenService tokenService;
 
 	@PostMapping("/login")
-	public ResponseEntity<JwtTokens> authorize(@Valid @RequestBody SignInRequest request) {
-		var authentication = authenticate(request.getEmail(), request.getPassword());
-		return ResponseEntity.ok().body(generateJwtTokens(request.getEmail(), authentication));
+	public ResponseEntity<TokenResponse> authorize(@Valid @RequestBody SignInRequest request) {
+		return ResponseEntity.ok().body(loginService.authenticateUser(request));
 	}
 
-	@PostMapping("/reissue")
-	public ResponseEntity<JwtTokens> reissue(@Valid @RequestBody RefreshTokenRequest request) {
-		String currentRefreshToken = request.getRefreshToken();
-
-		validateRefreshToken(currentRefreshToken);
-
-		String username = tokenProvider.getUsernameFromToken(currentRefreshToken);
-
-		var userDetails = loginService.loadUserByUsername(username);
-		var authentication = new UsernamePasswordAuthenticationToken(
-			userDetails, null, userDetails.getAuthorities());
-
-		return ResponseEntity.ok().body(generateJwtTokens(username, authentication));
-	}
-
-	private JwtTokens generateJwtTokens(String username, Authentication authentication) {
-		String accessToken = tokenProvider.createAccessToken(authentication);
-		String refreshToken = createAndSaveRefreshToken(username, authentication);
-
-		return new JwtTokens(accessToken, refreshToken);
-	}
-
-	private Authentication authenticate(String email, String password) {
-		var authenticationToken = new UsernamePasswordAuthenticationToken(email, password);
-		var authentication = authenticationManagerBuilder.getObject()
-			.authenticate(authenticationToken);
-		SecurityContextHolder.getContext().setAuthentication(authentication);
-		return authentication;
-	}
-
-	private String createAndSaveRefreshToken(String email, Authentication authentication) {
-		String refreshToken = tokenProvider.createRefreshToken(authentication);
-
-		refreshTokenService.deleteTokenByEmail(email);
-		refreshTokenService.save(email, refreshToken);
-
-		return refreshToken;
-	}
-
-	private void validateRefreshToken(String token) {
-		if (!refreshTokenService.validateToken(token)) {
-			throw new AuthException(ErrorCode.INVALID_TOKEN);
-		}
-	}
-
-	private record JwtTokens(@JsonProperty("access_token") String accessToken,
-							 @JsonProperty("refresh_token") String refreshToken) {
+	@PostMapping("/refresh")
+	public ResponseEntity<TokenResponse> refresh(@Valid @RequestBody RefreshTokenRequest request) {
+		return ResponseEntity.ok().body(tokenService.refreshJwtTokens(request));
 	}
 }
