+++ src/main/java/kr/binarybard/hireo/api/auth/dto/TokenResponse.java
@@ -1,0 +1,7 @@
+package kr.binarybard.hireo.api.auth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record TokenResponse(@JsonProperty("access_token") String accessToken,
+							@JsonProperty("refresh_token") String refreshToken) {
+}
